+++ CHANGELOG.md
@@ -4,10 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [1.13.0] - 2024-01-30
+# [1.13.0] - 2024-01-31
 - Upgrade `update` command: now includes support for modifications to API Gateway
 leveraging the OpenAPI v3 specification, allowing for seamless updates and more
 comprehensive management of API Gateway resources.
+
+# [1.12.1] - 2024-01-30
+- Fixed API Gateway deployment with the default value(300) of the `Cache time to live` parameter if the specified value is `0`
+
+# [1.12.0] - 2024-01-30
+### Added
+- The @LambdaHandler annotation for Java plugin improved to support the lambda 'architecture' management
+- The Java example 'java-layer-url' extended to use Lambda architecture management
+### Changed
+- The deployment-configuration-processor version bumped to 1.11.0
 
 # [1.11.0] - 2024-01-26
 - Implemented lambda function processor architecture type management